+++ package.json
@@ -21,7 +21,7 @@
     "react-mdl": "^2.1.0",
     "react-particles-js": "^3.4.1",
     "react-responsive": "^8.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.0.0-beta.8",
     "react-scripts": "4.0.3",
     "react-typed": "^1.2.0",
     "web-vitals": "^1.1.0"
