+++ src/Components/Experience/Experience.css
@@ -1,0 +1,41 @@
+.experience{
+    display: flex;
+    flex-direction: row;
+    background: white;
+    align-items: center;
+    justify-content: space-evenly;
+    margin: 10px;
+    padding: 20px;
+    max-height: 400px;
+    min-width: 100px;
+    z-index: 1;
+    width: 100%;
+}
+.experience__Companyimage {
+    max-height: 200px;
+    width: 100%;
+    object-fit: contain;
+    margin-bottom: 15px;
+}
+.experience__Info{
+    width: 100%;
+    height: 100px;
+    margin-bottom: 15px;
+}
+
+.experience__Title {
+    font-weight: 600;
+    font-size: 18px;
+}
+.experience__Companyname {
+    font-weight: 500;
+    font-size: 16px;
+}
+.experience__Experience {
+    font-weight: 500;
+    font-size: 16px;
+    color: darkgreen;
+}
+.experience__Projects {
+    font-size: 16px;
+}
