+++ src/Components/Navbar/Navbar.css
@@ -1,0 +1,34 @@
+.navbar{
+    height: 60px;
+    display: flex;
+    align-items: center;
+    background-color: #131921;
+    position: sticky;
+    top: 0;
+    z-index: 100;
+}
+
+.navbar__nav {
+    display: flex;
+    /* flex-direction: column; */
+    justify-content: space-evenly;
+    margin-left: 10px;
+    margin-right: 10px;
+}
+.navbar__Logo{   
+    width: 100px;
+    height: 60px;
+    object-fit: contain;
+    /* margin-top: 18px; */
+    /* margin: 0 20px; */
+}
+.navbar__Option {
+    color:white;
+    display: flex;
+    flex-direction: column;
+    margin-left: 10px;
+    margin-right: 10px;
+}
+.navbar__Option_lineOne{
+    font-size: 12px;
+}
