+++ src/Components/Portfolio/Portfolio.css
@@ -1,0 +1,23 @@
+.portfolio {
+    display: flex;
+    flex-direction: row;
+    background: white;
+    align-items: center;
+    justify-content: space-evenly;
+    margin: 10px;
+    padding: 20px;
+    max-height: 400px;
+    min-width: 100px;
+    /* z-index: -1; */
+    width: 100%;
+    position: sticky;
+}
+.portfolio__Image {
+    width: 50%;
+    object-fit: contain;
+}
+.portfolio__Info {
+    width: 100%;
+    height: 100px;
+    margin-bottom: 15px;
+}
